[update] - atualização na arquitetura do aplicativo para padrões modernos.
