+++ app/src/main/java/br/com/phmiranda/agenda/client/AlunoClient.java
@@ -1,0 +1,5 @@
+package br.com.phmiranda.agenda.client;
+
+public class AlunoClient {
+
+}
