+++ app/src/main/java/br/com/phmiranda/agenda/domain/dto/AlunoDto.java
@@ -1,0 +1,4 @@
+package br.com.phmiranda.agenda.domain.dto;
+
+public class AlunoDto {
+}
