+++ app/src/main/java/br/com/phmiranda/agenda/domain/enums/AlunoStatus.java
@@ -1,0 +1,4 @@
+package br.com.phmiranda.agenda.domain.enums;
+
+public enum AlunoStatus {
+}
