+++ app/src/main/java/br/com/phmiranda/agenda/service/AlunoService.java
@@ -1,0 +1,4 @@
+package br.com.phmiranda.agenda.service;
+
+public class AlunoService {
+}
