+++ app/src/main/java/br/com/phmiranda/agenda/view/ListaAlunosActivity.java
@@ -7,7 +7,6 @@ import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.ListView;
-import android.widget.Toast;
 
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
@@ -17,8 +16,8 @@ import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import java.util.List;
 
 import br.com.phmiranda.agenda.R;
-import br.com.phmiranda.agenda.entity.Aluno;
-import br.com.phmiranda.agenda.repository.AlunoDao;
+import br.com.phmiranda.agenda.domain.Aluno;
+import br.com.phmiranda.agenda.repository.dao.AlunoDao;
 
 public class ListaAlunosActivity extends AppCompatActivity {
